Fix timing issue for sentinel master-reboot test (#14312)
From the following logs, if we are in a slow environment, the election
process of sentinels may become very slow.
Even if the master instance that was restarted and is slowly loading RDB
has already been loaded, the election just gets started.

This PR makes the master load the RDB more slowly, and fixes the missing
of reseting reset `key-load-delay` for the master node.